Split background image plugin
